+++ Include/Acidanthera/Library/OcBootManagementLib.h
@@ -1099,6 +1099,10 @@ typedef struct OC_BOOT_ARGUMENTS_ {
   UINT32            *MemoryMapDescriptorSize;
   UINT32            *MemoryMapDescriptorVersion;
   CHAR8             *CommandLine;
+  UINT32            *KernelAddrP;
+  UINT32            *SystemTableP;
+  UINT32            *RuntimeServicesPG;
+  UINT64            *RuntimeServicesV;
   UINT32            *DeviceTreeP;
   UINT32            *DeviceTreeLength;
   UINT32            *CsrActiveConfig;