+++ Changelog.md
@@ -11,6 +11,7 @@ OpenCore Changelog
 - Updated builtin firmware versions for SMBIOS and the rest
 - Fixed ACPI patches not applying if tables are in locked memory
 - Fixed `EnableSafeModeSlide` on macOS 11.0
+- Added `AllowRelocationBlock` quirk for older macOS and safe mode
 
 #### v0.6.3
 - Added support for xml comments in plist files